# Reuse `HashSet` instances, reduce LINQ usage, avoid reallocating `StructureValue._properties`