… src/Serilog/Capturing/GetablePropertyFinder.cs
… test/Serilog.Tests/Capturing/GettablePropertyFinderTests.cs